requirements.txt
asgiref==3.8.1
Django==5.1
django-cors-headers==4.4.0
django-filter==24.3
django-rename-app==0.1.7
djangorestframework==3.15.2
et-xmlfile==1.1.0
mysqlclient==2.2.4
numpy==2.1.0
openpyxl==3.1.5
pandas==2.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1